requirements.txt
Flask-Limiter==2.5.0
requests==2.28.1
gunicorn==20.1.0
certifi
cryptography
pyOpenSSL
python-dotenv
fastapi~=0.79.0
uvicorn~=0.18.2
starlette==0.19.1
opencc-python-reimplemented
